refactor: extract @mtosity/admin package (auth, pages, API handlers)
Web app keeps thin /admin and /api/admin route files; NextAuth config moves
to @mtosity/admin/auth, consumed by middleware and the auth catch-all route.
