+++ apps/web/package.json
@@ -15,6 +15,7 @@
   },
   "dependencies": {
     "@huggingface/transformers": "^4.2.0",
+    "@mtosity/admin": "workspace:*",
     "@mtosity/design-system": "workspace:*",
     "@mtosity/lib": "workspace:*",
     "@mtosity/tool-img-grid": "workspace:*",
@@ -22,11 +23,6 @@
     "@mtosity/tool-speech-to-text": "workspace:*",
     "@lottiefiles/react-lottie-player": "^3.5.4",
     "@neondatabase/serverless": "^1.1.0",
-    "@tiptap/extension-image": "^3.26.0",
-    "@tiptap/extension-link": "^3.26.0",
-    "@tiptap/pm": "^3.26.0",
-    "@tiptap/react": "^3.26.0",
-    "@tiptap/starter-kit": "^3.26.0",
     "@types/prismjs": "^1.26.5",
     "@upstash/ratelimit": "^2.0.8",
     "@upstash/redis": "^1.37.0",
@@ -38,7 +34,6 @@
     "framer-motion": "catalog:",
     "lottie-react": "^2.4.1",
     "next": "catalog:",
-    "next-auth": "5.0.0-beta.31",
     "prismjs": "^1.30.0",
     "react": "catalog:",
     "react-dom": "catalog:",
