+++ packages/admin/src/api/note.ts
@@ -1,0 +1,64 @@
+import { NextResponse } from "next/server";
+import { auth } from "../auth";
+import {
+  deleteNote,
+  getNoteById,
+  updateNote,
+  type NoteInput,
+} from "@mtosity/lib/notes";
+
+
+type Ctx = { params: Promise<{ id: string }> };
+
+export async function GET(_req: Request, { params }: Ctx) {
+  const session = await auth();
+  if (!session) {
+    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  }
+  const { id } = await params;
+  const note = await getNoteById(id);
+  if (!note) {
+    return NextResponse.json({ error: "Not found" }, { status: 404 });
+  }
+  return NextResponse.json({ note });
+}
+
+export async function PUT(req: Request, { params }: Ctx) {
+  const session = await auth();
+  if (!session) {
+    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  }
+  const { id } = await params;
+  try {
+    const body = (await req.json()) as Partial<NoteInput>;
+    if (!body.title?.trim()) {
+      return NextResponse.json({ error: "Title is required" }, { status: 400 });
+    }
+    const note = await updateNote(id, {
+      title: body.title.trim(),
+      bodyHtml: body.bodyHtml ?? "",
+      bodyJson: body.bodyJson,
+      summary: body.summary,
+      published: body.published,
+    });
+    if (!note) {
+      return NextResponse.json({ error: "Not found" }, { status: 404 });
+    }
+    return NextResponse.json({ note });
+  } catch {
+    return NextResponse.json({ error: "Failed to update note" }, { status: 500 });
+  }
+}
+
+export async function DELETE(_req: Request, { params }: Ctx) {
+  const session = await auth();
+  if (!session) {
+    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  }
+  const { id } = await params;
+  const ok = await deleteNote(id);
+  if (!ok) {
+    return NextResponse.json({ error: "Not found" }, { status: 404 });
+  }
+  return NextResponse.json({ ok: true });
+}
