+++ packages/admin/src/api/upload.ts
@@ -1,0 +1,27 @@
+import { NextResponse } from "next/server";
+import { put } from "@vercel/blob";
+import { auth } from "../auth";
+
+
+// POST /api/admin/upload — push an image to Vercel Blob, return its public URL.
+export async function POST(req: Request) {
+  const session = await auth();
+  if (!session) {
+    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  }
+  try {
+    const form = await req.formData();
+    const file = form.get("file");
+    if (!(file instanceof File)) {
+      return NextResponse.json({ error: "No file provided" }, { status: 400 });
+    }
+    const safeName = file.name.replace(/[^a-zA-Z0-9._-]/g, "_");
+    const blob = await put(`notes/${Date.now()}-${safeName}`, file, {
+      access: "public",
+      addRandomSuffix: true,
+    });
+    return NextResponse.json({ url: blob.url });
+  } catch {
+    return NextResponse.json({ error: "Upload failed" }, { status: 500 });
+  }
+}
