+++ packages/admin/src/components/NoteEditor.tsx
