+++ packages/admin/src/index.ts
@@ -1,0 +1,6 @@
+export { default as AdminLayout } from "./pages/AdminLayout";
+export { default as AdminDashboard } from "./pages/AdminDashboard";
+export { default as AdminNotes } from "./pages/AdminNotes";
+export { default as NewNotePage } from "./pages/NewNotePage";
+export { default as EditNotePage } from "./pages/EditNotePage";
+export { default as AdminPhotography } from "./pages/AdminPhotography";
