+++ packages/admin/src/pages/NewNotePage.tsx
@@ -1,0 +1,6 @@
+import NoteEditor from "../components/NoteEditor";
+
+
+export default function NewNotePage() {
+  return <NoteEditor />;
+}
