+++ packages/admin/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@mtosity/typescript-config/react-library.json",
+  "include": ["src"],
+  "exclude": ["node_modules"]
+}
